+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
 
 ### Added
 
@@ -21,6 +21,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - An error is now thrown if mermaid-cli fails to render an SVG.
   Previously, the `Promise` returned by `renderMermaidFile` would stall forever
   if `mermaid-cli` exited without an error code, and without rendering an SVG.
+
+### Dependencies
+
+- Updated `husky` to v6.0.0 (v7.0.0 held back due to Node v10 support)
+- Limit `mermaid-cli` version to `<=8.10.1`.
+  This is because [mermaid-cli 8.10.2](https://github.com/mermaid-js/mermaid-cli/releases/tag/8.10.2)
+  upgrades to [puppeteer v10.0.0](https://github.com/mermaid-js/mermaid-cli/pull/128),
+  which drops Node v10 support.
 
 ## [1.0.2] - 2021-03-29